+++ lib/src/day_views/category/category_2d_day_view.dart
@@ -27,7 +27,7 @@ class CategoryDayView<T extends Object> extends StatelessWidget implements Calen
   @override
   Widget build(BuildContext context) {
     final eventPartLength = MediaQuery.sizeOf(context).width - config.timeColumnWidth;
-    final columnWidth = config.allowHorizontalScroll ? eventPartLength / config.columnsPerPage : eventPartLength / categories.length;
+    final columnWidth = config.allowHorizontalScroll ? eventPartLength / config.columnsPerPage : eventPartLength / config.columnsPerPage;
 
     controller?.calbliate(
       columnWidth,
@@ -38,7 +38,7 @@ class CategoryDayView<T extends Object> extends StatelessWidget implements Calen
       columnCount: categories.length + 1,
       rowCount: config.timeList.length + 1,
       pinnedColumnCount: 1,
-      pinnedRowCount: 1,
+      pinnedRowCount: config.freezeCategoryTitleRow ? 1 : 0,
       verticalDetails: ScrollableDetails(
         direction: AxisDirection.down,
         physics: const ClampingScrollPhysics(),